Set mailbox texture via reflection rather than command
We still don't have an API endpoint for directly setting player head textures,
but it's a whole lot easier now than it was in 2015 when I had to use vanilla
commands to spawn the mailboxes in. At least we can reflect into the SkullMeta
and set the profile field. I thought I'd surely be able to use the new
PlayerProfile class, but that doesn't seem to support direct setting of base64
textures.

* Add Mojang's AuthLib as a dependency.
* Replace spawning of mailboxes via commands with initialization via reflection.
* Include an enumeration of all possible mailbox colors.
* Replace mailbox color in-place on punch. No longer removing and then giving.